simle-agent/pyproject.toml
[project]
name = "simle-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]
